+++ packages/layers/src/scale-bar-layer.js
@@ -1,32 +1,31 @@
 import { CompositeLayer, COORDINATE_SYSTEM } from '@deck.gl/core';
 import { LineLayer, TextLayer } from '@deck.gl/layers';
-import { range, makeBoundingBox } from './utils';
+import { range, makeBoundingBox, snapValue, sizeToMeters } from './utils';
 
 import { DEFAULT_FONT_FAMILY } from '@vivjs/constants';
 
 function getPosition(boundingBox, position, length) {
-  const viewLength = boundingBox[2][0] - boundingBox[0][0];
+  const viewWidth = boundingBox[2][0] - boundingBox[0][0];
+  const viewHeight = boundingBox[2][1] - boundingBox[0][1];
   switch (position) {
     case 'bottom-right': {
-      const yCoord =
-        boundingBox[2][1] - (boundingBox[2][1] - boundingBox[0][1]) * length;
-      const xLeftCoord = boundingBox[2][0] - viewLength * length;
+      const yCoord = boundingBox[2][1] - viewHeight * length;
+      const xLeftCoord = boundingBox[2][0] - viewWidth * length;
       return [yCoord, xLeftCoord];
     }
     case 'top-right': {
-      const yCoord = (boundingBox[2][1] - boundingBox[0][1]) * length;
-      const xLeftCoord = boundingBox[2][0] - viewLength * length;
+      const yCoord = boundingBox[0][1] + viewHeight * length;
+      const xLeftCoord = boundingBox[2][0] - viewWidth * length;
       return [yCoord, xLeftCoord];
     }
     case 'top-left': {
-      const yCoord = (boundingBox[2][1] - boundingBox[0][1]) * length;
-      const xLeftCoord = viewLength * length;
+      const yCoord = boundingBox[0][1] + viewHeight * length;
+      const xLeftCoord = boundingBox[0][0] + viewWidth * length;
       return [yCoord, xLeftCoord];
     }
     case 'bottom-left': {
-      const yCoord =
-        boundingBox[2][1] - (boundingBox[2][1] - boundingBox[0][1]) * length;
-      const xLeftCoord = viewLength * length;
+      const yCoord = boundingBox[2][1] - viewHeight * length;
+      const xLeftCoord = boundingBox[0][0] + viewWidth * length;
       return [yCoord, xLeftCoord];
     }
     default: {
@@ -45,7 +44,8 @@ const defaultProps = {
   unit: { type: 'string', value: '', compare: true },
   size: { type: 'number', value: 1, compare: true },
   position: { type: 'string', value: 'bottom-right', compare: true },
-  length: { type: 'number', value: 0.085, compare: true }
+  length: { type: 'number', value: 0.085, compare: true },
+  snap: { type: 'boolean', value: false, compare: true }
 };
 /**
  * @typedef LayerProps
@@ -57,6 +57,7 @@ const defaultProps = {
  * @property {Array=} boundingBox Boudning box of the view in which this should render.
  * @property {string=} id Id from the parent layer.
  * @property {number=} length Value from 0 to 1 representing the portion of the view to be used for the length part of the scale bar.
+ * @property {boolean} snap If true, aligns the scale bar value to predefined intervals for clearer readings, adjusting units if necessary.
  */
 
 /**
@@ -65,7 +66,7 @@ const defaultProps = {
  */
 const ScaleBarLayer = class extends CompositeLayer {
   renderLayers() {
-    const { id, unit, size, position, viewState, length } = this.props;
+    const { id, unit, size, position, viewState, length, snap } = this.props;
     const boundingBox = makeBoundingBox(viewState);
     const { zoom } = viewState;
     const viewLength = boundingBox[2][0] - boundingBox[0][0];
@@ -76,15 +77,38 @@ const ScaleBarLayer = class extends CompositeLayer {
       2 ** (-zoom + 1.5),
       (boundingBox[2][1] - boundingBox[0][1]) * 0.007
     );
-    const numUnits = barLength * size;
+
+    // Initialize values for the non-snapped case.
+    let adjustedBarLength = barLength;
+    let displayNumber = (barLength * size).toPrecision(5);
+    let displayUnit = unit;
+    if (snap) {
+      // Convert `size` to meters, since `snapValue`
+      // assumes the value is in meters.
+      const meterSize = sizeToMeters(size, unit);
+      const numUnits = barLength * meterSize;
+      // Get snapped value in original units and new units.
+      const [snappedOrigUnits, snappedNewUnits, snappedUnitPrefix] =
+        snapValue(numUnits);
+      // We adjust the bar length by using the ratio of the snapped
+      // value in original units to the original value passed to `snapValue` (which is based on `meterSize`).
+      adjustedBarLength = snappedOrigUnits / meterSize;
+      displayNumber = snappedNewUnits;
+      displayUnit = `${snappedUnitPrefix}m`;
+    }
+
     const [yCoord, xLeftCoord] = getPosition(boundingBox, position, length);
+    const xRightCoord = xLeftCoord + barLength;
+
+    const isLeft = position.endsWith('-left');
+
     const lengthBar = new LineLayer({
       id: `scale-bar-length-${id}`,
       coordinateSystem: COORDINATE_SYSTEM.CARTESIAN,
       data: [
         [
-          [xLeftCoord, yCoord],
-          [xLeftCoord + barLength, yCoord]
+          [isLeft ? xLeftCoord : xRightCoord - adjustedBarLength, yCoord],
+          [isLeft ? xLeftCoord + adjustedBarLength : xRightCoord, yCoord]
         ]
       ],
       getSourcePosition: d => d[0],
@@ -97,8 +121,14 @@ const ScaleBarLayer = class extends CompositeLayer {
       coordinateSystem: COORDINATE_SYSTEM.CARTESIAN,
       data: [
         [
-          [xLeftCoord, yCoord - barHeight],
-          [xLeftCoord, yCoord + barHeight]
+          [
+            isLeft ? xLeftCoord : xRightCoord - adjustedBarLength,
+            yCoord - barHeight
+          ],
+          [
+            isLeft ? xLeftCoord : xRightCoord - adjustedBarLength,
+            yCoord + barHeight
+          ]
         ]
       ],
       getSourcePosition: d => d[0],
@@ -111,8 +141,14 @@ const ScaleBarLayer = class extends CompositeLayer {
       coordinateSystem: COORDINATE_SYSTEM.CARTESIAN,
       data: [
         [
-          [xLeftCoord + barLength, yCoord - barHeight],
-          [xLeftCoord + barLength, yCoord + barHeight]
+          [
+            isLeft ? xLeftCoord + adjustedBarLength : xRightCoord,
+            yCoord - barHeight
+          ],
+          [
+            isLeft ? xLeftCoord + adjustedBarLength : xRightCoord,
+            yCoord + barHeight
+          ]
         ]
       ],
       getSourcePosition: d => d[0],
@@ -125,8 +161,13 @@ const ScaleBarLayer = class extends CompositeLayer {
       coordinateSystem: COORDINATE_SYSTEM.CARTESIAN,
       data: [
         {
-          text: numUnits.toPrecision(5) + unit,
-          position: [xLeftCoord + barLength * 0.5, yCoord + barHeight * 4]
+          text: `${displayNumber}${displayUnit}`,
+          position: [
+            isLeft
+              ? xLeftCoord + barLength * 0.5
+              : xRightCoord - barLength * 0.5,
+            yCoord + barHeight * 4
+          ]
         }
       ],
       getColor: [220, 220, 220, 255],
@@ -135,7 +176,7 @@ const ScaleBarLayer = class extends CompositeLayer {
       sizeUnits: 'meters',
       sizeScale: 2 ** -zoom,
       characterSet: [
-        ...unit.split(''),
+        ...displayUnit.split(''),
         ...range(10).map(i => String(i)),
         '.',
         'e',