+++ packages/layers/tests/utils.spec.js
@@ -1,6 +1,11 @@
 import test from 'tape-catch';
 import { range } from '../src/multiscale-image-layer/utils';
-import { padWithDefault, padContrastLimits } from '../src/utils';
+import {
+  padWithDefault,
+  padContrastLimits,
+  snapValue,
+  sizeToMeters
+} from '../src/utils';
 import { createTestContext } from '@luma.gl/test-utils';
 import GL from '@luma.gl/constants';
 import { FEATURES } from '@luma.gl/webgl';
@@ -163,5 +168,60 @@ test('getRenderingAttrs WebGL2', t => {
   } catch (e) {
     t.fail(e);
   }
+  t.end();
+});
+
+test('sizeToMeters test', t => {
+  t.plan(5);
+  t.equal(
+    sizeToMeters(5, 'km'),
+    5000,
+    'Size in meters when unit is kilometers'
+  );
+  t.equal(
+    sizeToMeters(10, 'm'),
+    10,
+    'Size in meters when unit is already meters'
+  );
+  t.equal(
+    sizeToMeters(10, 'cm'),
+    0.1,
+    'Size in meters when unit is centimeters'
+  );
+  // These have floating point errors, so we check that
+  // they are within 1e-10 of the expected value.
+  t.ok(
+    sizeToMeters(10, 'µm') - 1e-5 < 1e-10,
+    'Size in meters when micrometers with Greek letter'
+  );
+  t.ok(
+    sizeToMeters(10, 'um') - 1e-5 < 1e-10,
+    'Size in meters when micrometers with regular u character'
+  );
+  t.end();
+});
+
+test('snapValue test', t => {
+  t.plan(4);
+  t.deepEqual(
+    snapValue(1.234),
+    [2, 2, ''],
+    'Snapping value inside extent of targets'
+  );
+  t.deepEqual(
+    snapValue(0.0234),
+    [0.025, 25, 'm'],
+    'Snapping value below minimum target'
+  );
+  t.deepEqual(
+    snapValue(2345.0),
+    [3000, 3, 'k'],
+    'Snapping value above maximum target'
+  );
+  t.deepEqual(
+    snapValue(999.0),
+    [1000, 1, 'k'],
+    'Snapping value between 500 and 1000'
+  );
   t.end();
 });