+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## In Progress
 
 ### Added
+- Added ScaleBarLayer snapping feature. Will convert scale value to one of `[1, 2, 3, 4, 5, 10, 20, 25, 50, 100, 200, 250, 500]` and adjust units accordingly, rescaling the length of the bar in pixels to match. Use `snap: true` prop to enable.
 
 ### Changed
 - Fixed typo in `sites/docs/src/IMAGE_RENDERING.md`